requirements.txt
langchain-chroma==0.1.4
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
python-dotenv==1.0.1
langchain-community
langchain-huggingface
pypdf
sentence-transformers>=3.0.0
huggingface_hub>=0.23.0
langchain-groq